requirements.txt
asgiref==3.5.2
cachetools==5.2.0
cbrf==0.5.0
certifi==2022.6.15
charset-normalizer==2.0.12
dbf-light==1.0.0
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
django-restframework==0.0.1
djangorestframework==3.13.1
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
greenlet==1.1.2
gspread==5.4.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
Markdown==3.3.7
numpy==1.23.0
oauth2client==4.1.3
oauthlib==3.2.0
protobuf==4.21.1
psycopg-binary==3.0.15
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycbrf==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
